admin
admin

+++ bee-admin/src/main/java/com/github/xhrg/bee/admin/dto/ResponseList.java
@@ -9,6 +9,5 @@ public class ResponseList {
 
     private List<?> items;
 
-    private int total;
-
+    private long total;
 }
